// blockwinz-FE/src/shared/enums/gameType.enum.ts
export enum GameTypeEnum {
    DiceGame = 'DiceGame',
    LimboGame = 'LimboGame',
    CoinFlipGame = 'CoinFlipGame',
    CrashGame = 'CrashGame',
    KenoGame = 'KenoGame',
    PlinkoGame = 'PlinkoGame',
    RouletteGame = 'RouletteGame',
    MinesGame = 'MinesGame',
    BlackjackGame = 'BlackjackGame',
    WheelGame = 'WheelGame',
    BaccaratGame = 'BaccaratGame',
    HiloGame = 'HiloGame',
}

export enum GameCategoryEnum {
    ALL = 'all',
    ORIGINALS = 'originals',
    MULTIPLAYER = 'multiplayer',
    SPORTS = 'sports',
}

export enum MultiplayerGameTypeEnum {
    TicTacToeGame = 'TicTacToeGame',
}